setup: script files & folders structures
