+++ checklist.html
@@ -30,7 +30,7 @@
           </div>
 
           <div id="right-side">
-            <button onclick="SwitchMode()" id="mode"><i class="ph ph-moon"></i></button>
+            <button id="mode"><i class="ph ph-moon"></i></button>
           </div>
         </nav>
       </div>
@@ -43,15 +43,15 @@
     </main>
     <footer>
       <div class="container">
-          <div class="footer-content">
-            <h3>BagyoAlerto</h3>
-            <p>Stay prepared, stay safe. Emergency preparedness for families & individuals in typhoon-prone areas</p>
-            <p>Privacy-first &middot; Offline-ready &middot; No account required</p>
-          </div>
+        <div class="footer-content">
+          <h3>BagyoAlerto</h3>
+          <p>Stay prepared, stay safe. Emergency preparedness for families & individuals in typhoon-prone areas</p>
+          <p>Privacy-first &middot; Offline-ready &middot; No account required</p>
         </div>
+      </div>
     </footer>
   </div>
+  <script src="./scripts/index.js" type="module"></script>
 </body>
 
-  <script src="./scripts/index.js"></script>
 </html>
